move helpers to "test_cli.__init__"
